backend/requirements.txt
fastapi
pydantic
yt-dlp
openai
python-dotenv
gunicorn
